fix(security): resolve 4 security vulnerabilities from code review
- SQL injection: Replace f-string SQL with parameterized queries in memory_store_v2
- Type error: Change cpu_time_limit_seconds from float to int in sandbox config
- Path traversal: Use Path.relative_to() instead of str.startswith() in file_ops
- Retry logic: Fix off-by-one error (attempt < max_retries -> <=) in planning exceptions

All fixes verified with tests, mypy, ruff, and bandit.

Ultraworked with [Sisyphus](https://github.com/code-yeongyu/oh-my-openagent)

+++ src/persona_agent/tools/sandbox.py
@@ -25,7 +25,7 @@ class SandboxConfig:
 
     # Time limits
     timeout_seconds: float = 30.0
-    cpu_time_limit_seconds: float = 10.0
+    cpu_time_limit_seconds: int = 10
 
     # Memory limits
     max_memory_mb: int = 128
